pyproject.toml
[project]
name = "tg-recipe-bot"
version = "0.1.0"
description = "Telegram bot that generates recipes using an LLM."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.0.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F"]
